requirements.txt
agate==1.6.3
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.8
altgraph==0.17.4
argcomplete==2.0.0
asttokens==3.0.0
attrs==22.2.0
Babel==2.10.3
beautifulsoup4==4.11.2
Brlapi==0.8.4
catboost==1.2.8
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
csvkit==1.0.7
cupshelpers==1.0
cycler==0.12.1
dbfread==2.0.7
dbus-python==1.3.2
debugpy==1.8.16
decorator==5.2.1
et-xmlfile==1.0.1
executing==2.2.0
feedparser==6.0.10
filelock==3.18.0
fonttools==4.59.1
fsspec==2025.7.0
graphviz==0.21
greenlet==2.0.2
html5lib==1.1
httplib2==0.20.4
idna==3.3
imbalanced-learn==0.14.0
imblearn==0.0
iniconfig==1.1.1
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jdcal==1.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
leather==0.3.4
lightgbm==4.6.0
llvmlite==0.44.0
louis==3.24.0
lxml==4.9.2
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
more-itertools==8.10.0
mpmath==1.3.0
musicbrainzngs==0.7.1
mutagen==1.46.0
narwhals==2.1.2
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
olefile==0.46
openpyxl==3.0.9
packaging==25.0
pandas==2.3.1
pandas-stubs==2.3.0.250703
parsedatetime==2.6
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pluggy==1.0.0+repack
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.45.2
Pygments==2.19.2
PyGObject==3.42.2
PyICU==2.10.2
pyinotify==0.9.6
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.0.9
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pytest==7.2.1
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-debianbts==4.0.1
python-slugify==4.0.0
pytimeparse==1.1.5
pytz==2025.2
pyxdg==0.28
pyzmq==27.0.1
quodlibet==4.5.0
reportbug==12.0.0
requests==2.28.1
scikit-learn==1.7.1
scipy==1.16.1
sgmllib3k==1.0.0
shap==0.48.0
six==1.16.0
slicer==0.0.8
soupsieve==2.3.2
SQLAlchemy==1.4.46
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
types-pytz==2025.2.0.20250809
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.3.6
urllib3==1.26.12
wcwidth==0.2.13
webencodings==0.5.1
xdg==5
xgboost==3.0.4
xlrd==1.2.0
pymupdf==1.26.7
